Improves background slideshow
Adds another visual effect for the slide show ( breathing ) as well as
a new way to render the list of images for the slide show. This new
rendering method randomizes a list of images and displays a certain
amount of them, thus changing the images seen on each page refresh.